requirements.txt
python-socketio==5.8.0
eventlet==0.33.0
dnspython==2.2.1